% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Big_LD.R
\name{Big_LD}
\alias{Big_LD}
\title{Estimation of LD block regions}
\usage{
Big_LD(geno, SNPinfo, CLQcut = 0.5, clstgap = 40000, leng = 200,
  subSegmSize = 1500)
}
\arguments{
\item{geno}{A data frame or matrix of additive genotype data, each column is additive genotype of each SNP.}

\item{SNPinfo}{A data frame or matrix of SNPs information.  1st column is rsID and 2nd column is bp position.}

\item{CLQcut}{A numeric value of threshold for the correlation value |r|, between 0 to 1.}

\item{clstgap}{An integer value to specifing the threshold of physical distance (bp) between two consecutive SNPs
which do not belong to the same clique, i.e., if a physical distance between two consecutive SNPs in a clique
greater than \code{clstgap}, then the algorithm split the cliques satisfying each
clique do not contain such consecutive SNPs}

\item{leng}{An integer value to specify the number of SNPs in a preceding and a following region
of each sub-region boundary, every SNP in a preceding and every SNP in a following region need to be in weak LD.}

\item{subSegmSize}{An integer value to specify the upper bound of the number of SNPs in a one-take sub-region.}
}
\value{
A data frame of block estimation result.
Each row of data frame shows the starting SNP and end SNP of each estimated LD block
}
\description{
\code{Big_LD} returns the estimation of LD block regions of given data.
}
\examples{

data(geno)
data(SNPinfo)
Big_LD(geno, SNPinfo)
Big_LD(geno, SNPinfo, CLQcut = 0.5, clstgap = 40000, leng = 200, subSegmSize = 1500)

}
\author{
Sun-Ah Kim <sunny03@snu.ac.kr>
}
\seealso{
\code{\link{CLQD}}, \code{\link{LDblockHeatmap}}
}

