% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLQD.R
\name{CLQD}
\alias{CLQD}
\title{partitioning into cliques}
\usage{
CLQD(subgeno, subSNPinfo, CLQcut = 0.5, clstgap = 40000,
  CLQmode = c("Density", "Maximal"), codechange = FALSE)
}
\arguments{
\item{subgeno}{A data frame or matrix of additive genotype data, each column is additive genotype of each SNP.}

\item{subSNPinfo}{A data frame or matrix of SNPs information. 1st column is rsID and 2nd column is bp position.}

\item{CLQcut}{A numeric value of threshold for the correlation value |r|, between 0 to 1.}

\item{clstgap}{An integer value to specifing the threshold of physical distance (bp) between two consecutive SNPs
which do not belong to the same clique, i.e., if a physical distance between two consecutive SNPs in a clique
greater than \code{clstgap}, then the algorithm split the cliques satisfying each
clique do not contain such consecutive SNPs}

\item{CLQmode}{A character string to specify the way to give priority among detected cliques.
if \code{CLQmode = "Density"} then the algorithm gives priority to the clique of largest value of \eqn{(Number of SNPs)/(range of clique)},
else if \code{CLQmode = "Maximal"}, then the algorithm gives priority to the largest clique.}

\item{codechange}{If \code{TRUE}, choose the cliques after code change procedure.}
}
\value{
A vector of cluster numbers of all SNPs (\code{NA} represents singleton cluster)
}
\description{
\code{CLQD} partitioning the given data into subgroups that contain SNPs which are highly correlated.
}
\examples{

data(geno)
data(SNPinfo)
CLQD(geno,SNPinfo,CLQcut = 0.5, clstgap= 40000, CLQmode = 'Maximal', codechange = FALSE)
CLQD(geno,SNPinfo,CLQcut = 0.5, clstgap= 40000, CLQmode = 'Density', codechange = FALSE)

}
\author{
Sun-Ah Kim <sunny03@snu.ac.kr>
}

