% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPART.R
\name{GPART}
\alias{GPART}
\title{Partitioning genodata based on the result obtained by using Big-LD and gene region information.}
\usage{
GPART(geno, SNPinfo, chrN, allgenelist, BigLDresult = NULL, minsize = 4,
  maxsize = 50)
}
\arguments{
\item{geno}{A data frame or matrix of additive genotype data, each column is additive genotype of each SNP.}

\item{SNPinfo}{A data frame or matrix of SNPs information.  1st column is rsID and 2nd column is bp position.}

\item{chrN}{A integer value to specify chromosome number of the given data.}

\item{allgenelist}{A data frame or matrix of Gene info data.
(1st col : Genename, 2nd col : chromosome, 3rd col : start bp,  4th col : end bp)}

\item{BigLDresult}{A data frame obtained by \code{Big_LD} function.
If \code{NULL}(default), the \code{GPART} function first excute \code{Big_LD} function to obtain LD blocks estimation result.}

\item{minsize}{A integer value specifying the lower bound of number of SNPs in a partition.}

\item{maxsize}{A integer value specifying the upper bound of number of SNPs in a partition.}
}
\value{

}
\description{
\code{GPART} partition the given genodata using the result obtained by Big-LD and gene region information.
The algorithm partition the whole sequence into sub sequences of which size do not exceed the given threshold.
}
\examples{
data(geno)
data(SNPinfo)
data(allgenelist)
GPART(geno, SNPinfo, 22, genelist)

}
\author{
Sun Ah Kim <sunny03@snu.ac.kr>, Yun Joo Yoo <yyoo@snu.ac.kr>
}
\seealso{
\code{\link{Big_LD}}
}

