% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDblockHeatmap.R
\name{LDblockHeatmap}
\alias{LDblockHeatmap}
\title{Draw LDblock results on LD heatmap.}
\usage{
LDblockHeatmap(geno, SNPinfo, chrN, showSNPs = NULL, LDblockResult = NULL,
  tick = c("bp", "rsID"), st.bp = 0, ed.bp = Inf, showLDsize = 3,
  savefile = FALSE)
}
\arguments{
\item{geno}{A data frame or matrix of additive genotype data, each column is additive genotype of each SNP.}

\item{SNPinfo}{A data frame or matrix of SNPs information.  1st column is rsID and 2nd column is bp position.}

\item{chrN}{A integer value to specify chromosome number of the given data.}

\item{showSNPs}{A data frame which is part of \code{SNPinfo} that you want to show in the result LDblock heatmap.
The default is \code{NULL}}

\item{LDblockResult}{A data frame obtained by \code{Big_LD} function.
If \code{NULL}(default), the \code{GPART} function first excute \code{Big_LD} function to obtain LD blocks estimation result.}

\item{tick}{A character string to specify how to show first SNPs and last SNPs of LD blocks,
in \code{"bp"} or in \code{"rsID"}.}

\item{st.bp}{A integer value to specify starting bp position of the region to draw.}

\item{ed.bp}{A integer value to specify end bp position of the region to draw.}

\item{showLDsize}{A integer value to specify the size (number of SNPs) of LDblocks to show ticks.
if \code{showLDsize = \eqn{k}}, then blocks of size equal or greater than k is shown with boundaries and
the "rsID" or "bp" of the first and last SNPs, however,
the blocks whose size is less than \eqn{k} is shown with only boundaries.}

\item{savefile}{logical. If \code{TRUE}, save tif file into work directory and
the file is named after the chromosome and the physical range to draw. The default is \code{FALSE}}
}
\value{
A grid graphical object of LD block heatmap.
The LD block heatmap will be presented on the screen after execution of the function.
#'
}
\description{
\code{LDblockHeatmap} shows the LDblock regions obtained by Big-LD algorithm.
}
\examples{

data(geno)
data(SNPinfo)
LDblockHeatmap(geno, SNPinfo,chrN = 22, showSNPs = NULL)
LDblockHeatmap(geno, SNPinfo, 22, showSNPs = SNPinfo[c(100, 200), ], showLDsize = 10, savefile = TRUE)
}
\author{
Sun-Ah Kim <sunny03@snu.ac.kr>, Yun Joo Yoo <yyoo@snu.ac.kr>
}
\seealso{
\code{\link{Big_LD}}
}

